Move abstract model classes to Models namespace.
